ais_receiver_faust/requirements.txt
libais==0.17
kafka-python
